+++ src/Games/capital/screens/capital.jsx
@@ -1,0 +1,3 @@
+export default function Capital() {
+  return <div>Work in progress..</div>;
+}