+++ src/screens/selectGame.jsx
@@ -10,32 +10,44 @@ export default function SelectGame() {
           <Title margin="2" text="Select the game !" />
         </RubberBand>
         <div
-          className="mt-2 w-full h-full sm:w-4/5 xl:w-3/5 border-2 rounded-lg lg:rounded-3xl
-      md:h-96 grid grid-rows-4 md:grid-cols-2 md:grid-rows-2 items-center dark:border-darkInput"
+          className="mt-2 w-full h-full sm:w-4/5 gap-4 rounded-lg lg:rounded-3xl
+      md:h-96 grid grid-rows-6 md:grid-cols-3 md:grid-rows-2 items-center dark:border-darkInput"
         >
           <SelectGameCard
             icon="olympic-games"
             text="Guess which country the flag belongs to"
-            border="border-0 border-b-2 md:border-r-2 dark:border-darkInput"
+            border="border-2 dark:border-darkInput"
             game="flag"
           ></SelectGameCard>
           <SelectGameCard
             icon="search"
-            border="border-0 border-b-2 dark:border-darkInput"
+            border="border-2 dark:border-darkInput"
             text="Guess the country from the clues given to you"
             game="investigation"
           ></SelectGameCard>
           <SelectGameCard
             icon="people"
-            border="border-0 border-b-2 md:border-b-0 md:border-r-2 dark:border-darkInput"
+            border="border-2 dark:border-darkInput"
             text="Guess which of the countries has the largest population"
             game="population"
           ></SelectGameCard>
           <SelectGameCard
             icon="australia"
-            border="border-0 dark:border-darkInput"
+            border="border-2 dark:border-darkInput"
             text="Find a country by crossing the borders"
             game="borders"
+          ></SelectGameCard>
+          <SelectGameCard
+            icon="eiffel-tower"
+            border="border-2 dark:border-darkInput"
+            text="Find a country by its capital"
+            game="capital"
+          ></SelectGameCard>
+          <SelectGameCard
+            icon="letter"
+            border="border-2 dark:border-darkInput"
+            text="Find as many countries as possible from a given first letter"
+            game="findletter"
           ></SelectGameCard>
         </div>
       </div>